// apps/web/src/lib/api.generated.d.ts
export interface components {
    schemas: {
// ...
        BoardOutput: {
            /** Version */
            version: number;
            /**
             * Status
             * @enum {string}
             */
            status: "idle" | "queued" | "running" | "ready" | "failed";
            /** Error Code */
            error_code: string | null;
            /** Summary */
            summary: components["schemas"]["SummaryOutput"][];
            /** Cards */
            cards: components["schemas"]["Card"][];
            /** Progress */
            progress: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}